% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{clusterAPAheatmap}
\alias{clusterAPAheatmap}
\title{Heatmap of cluster-specific APA genes}
\usage{
clusterAPAheatmap(ECoeffSig_Mat, FDR_P_cutoff = 0.05, CoeffCutoff = log(2))
}
\arguments{
\item{ECoeffSig_Mat}{Matrix containing estimated coefficients, unadjusted P values of Wald tests on coefficients, and SE returned by model mode \code{estimateSig}.}

\item{FDR_P_cutoff}{The cutoff for FDR-controlled P values of Wald tests. Default to 0.05. Only cluster-specific APA events with FDR-controlled P values of Wald test smaller than this number will be considered as 3'UTR lengthening or shortening.}

\item{CoeffCutoff}{The cutoff for estimated coefficients of logistic regression. Only cluster-specific APA events with absoluste value of estimated coeffcients greater than this cutoff will be considered as 3'UTR lengthening or shortening.}
}
\value{
\code{clusterAPAheatmap} returns a heatmap of cluster-specific APA genes
}
\description{
\code{clusterAPAheatmap} draws heatmap of cluster-specific APA genes to show the degree and direction of APA. This function first selects genes that are significant cluster-specific APA in any cluster. Then, it draws heatmap with estimated coefficients from logistic regression.
}
\examples{
clusterAPAheatmap(ECoeffSig_Mat = result_from_estimateSig$ECoeffSig_Mat, FDR_P_cutoff=0.05, CoeffCutoff=log(2))

}
