% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSig.R
\name{estimateSig}
\alias{estimateSig}
\title{Estimate significance of APA event at gene level}
\usage{
estimateSig(ISOMatrix, mode = "model", FDR_P_cutoff = 0.05)
}
\arguments{
\item{mode}{Character parameter. Either "test" or "model", indicating which approach should be used to estimate the significance of APA events. Default to model-based approach. Default to "model".}

\item{FDR_P_cutoff}{The cutoff for FDR-controlled P values. Default to 0.05.}

\item{IsoMatrix}{Isoform-specific count matrix returned by \code{readin}.}
}
\value{
Model-based mode will return a list object containing three elements. 1) siglist_FDRp contains significant APA event IDs and their gene-level FDR controlled P values. 2) ECoeff_Mat contains estimated coefficients, P values, and SE from logsitic regression model for all genes. 3) ECoeffSig_Mat contains estimated coefficients, P values, and SE from logsitic regression model for significant APA events only and will be used for visualization. \cr \cr Test-based mode will return a list of IDs of significant APA events.
}
\description{
\code{estimateSig} estimates significance of APA event at gene level using either model-based approach or test-based approach. Model-based approach will fit a logistic regression model for each gene and estimate significance using Likelihood ratio test. Test-based approach will condunct Fisher's exact test on all possible pairs of clusters (i.e. 1 vs. 2; 1 vs. 3; 2 vs. 3 if 3 clusters) for each gene and use FDR to adjust for multiple-testing issue. Test-based approach defines significant APA genes if they pass significance test in any pair of the clusters.
}
\examples{
result_list  <- estimateSig(ISOMatrix=Matix_returned_from_readin)
estimateSig(ISOMatrix=Matrix , mode = "model", FDR_P_cutoff=0.05)
estimateSig(ISOMatrix=Matrix , mode = "test", FDR_P_cutoff=0.05)

}
