% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_filter.R
\name{readin}
\alias{readin}
\title{Read in DaPars outputs and filter out low expressed genes.}
\usage{
readin(path, NAcutoff, CPMcutoff = 10)
}
\arguments{
\item{path}{Path to the folder containing (only) outputs from DaPars for all samples.}

\item{NAcutoff}{The number of NAs could be tolerated for each gene. \code{readin} will only keep genes with number of NAs <= NAcutoff.}

\item{CPMcutoff}{The cutoff for gene-wise sum of CPM. \code{readin} will calculate CPM for each long/short Isoform from each sample separately by CPM=(raw count)*10^6/lib_size. Then, only genes with rowSums(CPM)>CPMcutoff will be kept.}
}
\value{
The returns will be an isoform specific count matrix based on the DaPars estimation.
}
\description{
\code{readin} returns filtered isoform count matrix for APA identification.
}
\examples{
ISOMatrix  <- readin(path=getwd(),NAcutoff=3)
ISOMatrix  <- readin(path="Directory/storing/result/tables/from/previous/step",NAcutoff=3,CPMcutoff=10)


}
